chore(e2e): added e2e for unloop mechanism (#3529)
* fixed action btn spacing for loan tab

* added translations for loans tab

* added e2e

* distributed loans e2e in different files

* fixed failing e2e

* file renamed
